+++ pyramid_urireferencer/referencer.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import abc
+import six
 import requests
 
 import logging
@@ -9,17 +10,28 @@ from .models import RegistryResponse
 log = logging.getLogger(__name__)
 
 
+@six.add_metaclass(abc.ABCMeta)
 class AbstractReferencer:
     """
     This is an abstract class that defines what a Referencer needs to be able to handle.
 
     It does two things:
-
+    
         * Check if a uri is being used in this application and report on this.
         * Check if a certain uri is being used in another application by query
           a central registry.
+          * this requires a function :meth:`get_uri` to determine the uri of the current request
     """
-    __metaclass__ = abc.ABCMeta
+
+    @abc.abstractmethod
+    def get_uri(self, request):
+        """
+        This method extracts a uri from the request. This is the uri that needs to be checked.
+
+        :param request: :class:`pyramid.request.Request` with useful configuration information and connections
+                        of the application (registry, route_url, session) to determine the references
+        :rtype: string uri: URI of the resource we need to check for
+        """
 
     @abc.abstractmethod
     def references(self, uri, request):
@@ -44,12 +56,12 @@ class AbstractReferencer:
         """
 
 
+@six.add_metaclass(abc.ABCMeta)
 class Referencer(AbstractReferencer):
     """
     This is an implementation of the :class:`AbstractReferencer` that adds a
-    generic :meth:`is_referenced` method and a plain :meth:`references` method..
+    generic :meth:`is_referenced` method and plain methods: :meth:`references` and :meth:`get_uri` 
     """
-    __metaclass__ = abc.ABCMeta
 
     def __init__(self, registry_url, **kwargs):
         """